% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_utils.R
\name{impute_mean}
\alias{impute_mean}
\title{Impute mean for each numeric column in a data.frame}
\usage{
impute_mean(df)
}
\arguments{
\item{df}{A data.frame with possible missing values}
}
\value{
A data.frame with missing numeric values replaced by the column mean
}
\description{
Impute mean for each numeric column in a data.frame
}
