% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_utils.R
\name{impute_median}
\alias{impute_median}
\title{Impute median for numeric columns.}
\usage{
impute_median(df, train_frac = 1, seed = 1)
}
\arguments{
\item{df}{A data frame.}

\item{train_frac}{Fraction of rows to use when computing the mean.}

\item{seed}{RNG seed for reproducible sampling.}
}
\value{
A data frame with missing numeric values replaced by the median.
}
\description{
Analogous to `impute_mean`, but uses the median instead of the mean.
}
