% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_utils.R
\name{impute_mode}
\alias{impute_mode}
\title{Impute mode for columns (numeric or character)}
\usage{
impute_mode(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Impute mode for columns (numeric or character)
}
