% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_utils.R
\name{missing_mask}
\alias{missing_mask}
\title{Determine which elements of a vector are missing.}
\usage{
missing_mask(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical vector indicating missingness.
}
\description{
This helper treats `NA`, empty strings (`""`), and the literal
string `"NA"` as missing values.  For character and factor
variables, these values are considered missing; for other types, only
`NA` values are treated as missing.  The result is a logical vector
the same length as `x`.
}
