% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_utils.R
\name{needs_imputation}
\alias{needs_imputation}
\title{Does a vector contain any missing values?}
\usage{
needs_imputation(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Logical scalar indicating the presence of missing values.
}
\description{
Returns `TRUE` if any elements of `x` are considered missing
according to `missing_mask()`.  This helper is used to decide
whether a column needs imputation.
}
