% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_utils.R
\name{impute_median}
\alias{impute_median}
\title{Impute median for numeric columns}
\usage{
impute_median(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Impute median for numeric columns
}
